+++ spec/app/models/pack.rb
@@ -1,0 +1,3 @@
+class Pack
+  include Mongoid::Document
+end