+++ spec/app/models/subscription.rb
@@ -1,0 +1,4 @@
+class Subscription
+  include Mongoid::Document
+  has_many :packs, class_name: "ShippingPack"
+end