+++ CHANGELOG.md
@@ -146,6 +146,9 @@ For instructions on upgrading to newer versions, visit
 
 * \#2979 `pluck` no longer modifies the context in place. (Brian Goff)
 
+* \#2970 Fixed counter cache to properly use the name of the relation if available
+  then the inverse class name second if not.
+
 * \#2959 Nested attributes will now respect `autosave: false` if defined on the
   relation.